Make CLI functions callable from Python
